gh-422: Publish cjs and es modules to npm
