+++ lib/cjs/shims/document.js
@@ -1,0 +1,8 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
+var _default = document;
+exports.default = _default;
