+++ lib/es/index.js
@@ -1,0 +1,3 @@
+import Carousel from './components/Carousel';
+import Thumbs from './components/Thumbs';
+export { Carousel, Thumbs };
