+++ lib/es/shims/document.js
@@ -1,0 +1,1 @@
+export default document;
