--- lib/index.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="UTF-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-		<link rel="stylesheet" href="main.css"/>
-		<link rel="stylesheet" href="carousel.css"/>
-		<link href='http://fonts.googleapis.com/css?family=Roboto:100,700,400' rel='stylesheet' type='text/css'>
-	</head>
-	<body>
-    <!-- Google Analytics -->
-    <script>
-      window.ga=window.ga||function(){(ga.q=ga.q||[]).push(arguments)};ga.l=+new Date;
-      ga('create', 'UA-62004510-2', 'auto');
-      ga('send', 'pageview');
-    </script>
-    <script async src='https://www.google-analytics.com/analytics.js'></script>
-    <!-- End Google Analytics -->
-
-		<a href="https://github.com/leandrowd/react-responsive-carousel"><img style="position: absolute; top: 0; right: 0; border: 0;" src="https://camo.githubusercontent.com/365986a132ccd6a44c23a9169022c0b5c890c387/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f7265645f6161303030302e706e67" alt="Fork me on GitHub" data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_red_aa0000.png"></a>
-		<div class="wrapper">
-			<h1>React Responsive Carousel</h1>
-
-			<a href="https://badge.fury.io/js/react-responsive-carousel"><img src="https://badge.fury.io/js/react-responsive-carousel.svg" alt="npm version" height="18"></a>
-
-			<a href="https://travis-ci.org/leandrowd/react-responsive-carousel"><img src="https://travis-ci.org/leandrowd/react-responsive-carousel.svg?branch=master"></a>
-
-      <h2>Storybook</h2>
-      <p>Check it out these <a href="/storybook">cool demos</a> created using <a href="https://getstorybook.io/">storybook</a>. The source code for each example is available <a href="https://github.com/leandrowd/react-responsive-carousel/blob/master/stories/index.js">here</a>.
-
-      <h2>Summary</h2>
-			<ul class="summary">
-				<li><a href="#install">Installing</a></li>
-                <li><a href="#demos">Demo</a>
-				<li><a href="#contribute">Contributing</a></li>
-			</ul>
-
-			<h2 id="install">Installing</h2>
-			<code>
-				npm install react-responsive-carousel --save
-			</code>
-
-			<h2 id="demos">Demo</h2>
-			<section class="demos">
-				<article class="demo">
-					<h3 id="carousel">Carousel</h3>
-					<div class="example demo-carousel"></div>
-					<h5>Code:</h5>
-					<script src="https://gist.github.com/leandrowd/b8a9385b0b4b03054dc5.js"></script>
-				</article>
-			</section>
-
-			<h2 id="contributing">Contributing</h2>
-			<p>Please, feel free to contributing. You may file an issue or submit a pull request!</p>
-		</div>
-		<script src="main.js"></script>
-	</body>
-</html>
