--- lib/styles/main.css
@@ -1,101 +1,0 @@
-/********************************************
-	BREAKPOINT WIDTHS
-********************************************/
-/********************************************
-	FONTS
-********************************************/
-/********************************************
-	COLOURS
-********************************************/
-* {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  -webkit-box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  box-sizing: border-box; }
-
-body {
-  font-family: "Roboto", sans-serif;
-  font-size: 16px;
-  font-weight: 400;
-  color: #333; }
-
-h1, h2, h3, h4 {
-  font-weight: 700;
-  margin: 20px 0; }
-
-h1 {
-  font-weight: 700;
-  font-size: 2.5em; }
-
-h2 {
-  border-top: 1px solid #999;
-  padding-top: 40px;
-  margin-top: 40px;
-  font-size: 2em; }
-
-h3 {
-  font-size: 1.4em; }
-
-h4 {
-  font-size: 1.1em;
-  margin-bottom: 10px; }
-
-code {
-  border-radius: 20px;
-  background: #f5f5f5;
-  padding: 20px;
-  text-indent: 0;
-  color: #333;
-  display: block;
-  margin: auto;
-  overflow: scroll;
-  width: 100%; }
-
-.wrapper {
-  padding: 30px; }
-  .wrapper .demo-gallery, .wrapper .demo-slider, .wrapper .demo-carousel {
-    margin: 30px -30px; }
-  .wrapper .demo {
-    display: block;
-    margin: 30px 0;
-    padding-bottom: 30px;
-    border-bottom: 1px solid #333; }
-    .wrapper .demo:last-child {
-      border-bottom: 0; }
-
-.list-items, .summary {
-  margin-left: 50px; }
-  .list-items li, .summary li {
-    line-height: 2;
-    list-style-type: circle; }
-    .list-items li ul, .summary li ul {
-      margin-left: 30px; }
-
-.demo-gallery, .demo-slider, .demo-carousel {
-  margin: 50px auto; }
-
-.demo-gallery .thumb {
-  width: 200px; }
-
-.demo-carousel {
-  width: 100%;
-  margin: 30px auto !important; }
-  .demo-carousel .thumb {
-    width: 120px; }
-
-.example {
-  margin: 20px 0; }
-
-.parent {
-  background: #333; }
-
-.page-carousel {
-  padding: 40px;
-  border: 1px solid #fff;
-  width: 80%;
-  margin: 40px auto; }
-
-h5 {
-  margin-bottom: 10px; }
