+++ tasks/configs.js
@@ -1,24 +1,9 @@
-var sourceFolder = 'src';
-// development folder
-var devFolder = 'dev';
-// google pages folder
-var productionFolder = 'dist';
-// npm package folder
-var packageFolder = 'lib';
+const demoWebsite = 'temp/website';
+const npmPackage = 'lib';
+const source = 'src';
 
-var configs = {
-    folders: {
-        source: sourceFolder,
-        development: devFolder,
-        production: productionFolder,
-        package: packageFolder,
-    },
-    paths: {
-        source: './' + sourceFolder,
-        development: './' + devFolder,
-        production: './' + productionFolder,
-        package: './' + packageFolder,
-    },
+module.exports = {
+    demoWebsite,
+    npmPackage,
+    source,
 };
-
-module.exports = configs;
