+++ tasks/gh-pages.js
@@ -1,8 +1,9 @@
 var path = require('path');
 var ghPages = require('gh-pages');
+var configs = require('./configs');
 
 module.exports = function(done) {
-    return ghPages.publish(path.join(__dirname, '../temp/website'), function(err) {
+    return ghPages.publish(path.join(__dirname, `../${configs.demoWebsite}`), function(err) {
         if (err) {
             console.error('Failed to publish', err);
         } else {
